// frontend/src/api/generated/schema.ts
export interface components {
    schemas: {
// ...
        AnalysisJobRequest: {
            store_profile: components["schemas"]["StoreProfile"];
            research_request: components["schemas"]["ResearchRequest"];
            /** Official Data Requests */
            official_data_requests?: components["schemas"]["OfficialDataRequest"][];
        };
// ...
        OfficialDataRequest: {
            /** Provider */
            provider: string;
            /** Indicator Id */
            indicator_id: string;
            /** Request Params */
            request_params?: {
                [key: string]: string | number;
            };
            /**
             * Required
             * @default false
             */
            required: boolean;
            /** Max Age Days */
            max_age_days?: number | null;
            /** @default MONTHLY */
            target_frequency: components["schemas"]["ObservationFrequency"];
            /**
             * Transform
             * @default LAST
             */
            transform: string;
        };
// ...
        WhatIfRequest: {
            /**
             * Scenario Name
             * @default WHAT_IF
             */
            scenario_name: string;
            /**
             * Revenue Multiplier
             * @default 1
             */
            revenue_multiplier: number | string;
            /**
             * Variable Cost Multiplier
             * @default 1
             */
            variable_cost_multiplier: number | string;
            /**
             * Fixed Cost Multiplier
             * @default 1
             */
            fixed_cost_multiplier: number | string;
            /**
             * Interest Rate Delta
             * @default 0
             */
            interest_rate_delta: number | string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}